Upgrade dependency bounds, enable tasty-1.5.*
This is a new variant of https://github.com/stoeffel/tasty-test-reporter/pull/12 which does not add a cabal file and also bumps a few more upper bounds to suit stackage LTS-23.9 (`ghc-9.8.4`).

Would appreciate if you could upload a new revision on hackage, too, but it already helps if this gets onto master here.

This is a nice and useful tool that we would not want to miss in our build, thanks for putting it together!